Update kernel startup, fix incomming issues with revamping of task

+++ include/kernel/core.h
@@ -100,6 +100,6 @@ void arch_start();
 /* - */
 void module_loader();
 /* - */
-int task_start(const char *name, void *func, void *arg);
+size_t task_start(const char *name, void *func, void *arg);
 
 #endif  /* _KERNEL_CORE_H */
